task & styling completed. WIP: JSDoc annotations

+++ src/components/Library.js
@@ -14,21 +14,39 @@ class Library extends React.Component {
     };
 
     __addBook(book, target) {
-        console.log("Adding: " + book.title + " to " + target)
+        console.log("Adding: " + book.title + " to " + target);
         const bookList = this.state[target];
+
+        const allBooks = [...this.state.currentBooks, ...this.state.toReadBooks, ...this.state.readBooks];
+        
+        // Check if the book already exists in any list.
+        const isBookExists = allBooks.some((listBook) => listBook.key === book.key);
+        if (isBookExists) {
+            console.log("Book already exists in a list. Preventing addition..");
+            return; // Exit the function if the book already exists
+        }
+
+        // Push the book to the bookList
         bookList.push(book);
-        this.setState({ [target]: bookList })
-        localStorage.setItem(target, JSON.stringify(bookList))
+
+        // Update the state and localStorage
+        this.setState({ [target]: bookList });
+        localStorage.setItem(target, JSON.stringify(bookList));
     }
 
     __removeBook(book, source) {
-        console.log("Removing: " + book.title + " from " + source)
-        const bookList = this.state[source];
+        return new Promise((resolve, reject) => {
+            console.log("Removing: " + book.title + " from " + source)
+            const bookList = this.state[source];
+    
+            const newBookList = bookList.filter(listBook => (this.__filterID(listBook, book)))
+    
+            this.setState({ [source]: newBookList }, () => {
+                localStorage.setItem(source, JSON.stringify(newBookList));
+                resolve(); // Resolve the promise when the state is updated
+            });
 
-        const newBookList = bookList.filter(listBook => (this.__filterID(listBook, book)))
-
-        this.setState({ [source]: newBookList })
-        localStorage.setItem(source, JSON.stringify(newBookList))
+        })
     }
 
     __filterID = (listBook, book) => {
@@ -36,8 +54,15 @@ class Library extends React.Component {
     }
 
     moveBook = (book, target, source) => {
-        if (source !== null) this.__removeBook(book, source);
-        if (target !== null) this.__addBook(book, target);
+        if (source !== null) {
+          this.__removeBook(book, source).then(() => {
+            if (target !== null) {
+              this.__addBook(book, target);
+            }
+          });
+        } else if (target !== null) {
+          this.__addBook(book, target);
+        }
     }
 
     render() {
@@ -46,7 +71,7 @@ class Library extends React.Component {
                 <CurrentlyReading currentBooks={this.state.currentBooks} move={this.moveBook} />
                 <ToRead toReadBooks={this.state.toReadBooks} move={this.moveBook} />
                 <Read readBooks={this.state.readBooks} move={this.moveBook} />
-                <SearchBook addBookToCurrent={this.moveBook} />
+                <SearchBook move={this.moveBook} />
             </>
         )
     }
